Recoded EmbedBuilderExtension and added a few error handlings

+++ Kaida/Kaida.Database/Models/Users/User.cs
@@ -9,7 +9,7 @@ namespace Kaida.Database.Models.Users
         public ulong Id { get; set; }
         public DateTimeOffset? Birthdate { get; set; }
         public string Description { get; set; }
-        public int? SteamId { get; set; }
+        public ulong? SteamId { get; set; }
         public IList<Infraction> Infractions { get; set; }
     }
 }
